Fixed one problem in stage and added one new feature.  The problem was that when staging was not possible (e.g., running on a noma with full /scratch), then the copying of files into xrootd (the "secondary" target for a stageOut file) was not performed.  Now, the stageOut map is filled even when no staging is actually done and when finish() is called, the secondary target (typically xrootd) gets its copy.
The new feature is a function called dumpFileList(filename) which
dumps the actual (final) file names of all anticipated output files.
The use is to produce this during a pipeline line job and store the
file in the working directory; if the stream is rolled back, this file
can be read and the associated output files cleaned up.